some functions and fixes [done] (#2178)
* Add gmClassic_801B3E44 function

* Rename function and add new struct definition

* Refactor gmClassic_801B3E44 function and update references

* Change UnkFlagStruct to bit fields in gmclassic.h

* Update gmclassic.h

* Update gmclassic.c

* Update forward.h

* Update types.h

* Update gmclassic.h

* Update gmclassic.h

* Update gmclassic.h

* Update types.h

* Update forward.h

* Update types.h

* Update gm_17C0.h

* Update gm_17C0.c

* Update gm_17C0.c

* Update gmclassic.c

* Update gmclassic.h

* Rename `gmClassic` symbols

* Update gm_17C0.c

fn_8017DD7C match

* Update gm_17C0.c

* Update gm_17C0.h

* Update gm_17C0.c

* Update types.h

* Update gm_17C0.c

gm_8017E440

* Update gm_17C0.c

gm_8017E4C4

* Update gm_17C0.c

* Update gm_17C0.c

* Update gm_17C0.h

* Update gm_17C0.c

* Update gm_17C0.h

* Update gm_17C0.c

* Update gm_17C0.h

* Update gm_17C0.c

* Update types.h

* Update gm_17C0.c

* Update gm_17C0.c

* Fix callback signature

* Run `clang-format`

* Fix broken match

* Fix field doc comment

* Update lbarq.h

* Update lbarq.c

* Update lbarq.c

* Update it_8029DD58

* Update itlinkbomb.h

* Update it_27CF.h

* Update ftLk_SpecialLw.c

* itTomato

Refactor item destruction logic for itTomato.

* Refactor itTomato_ItemVars structure

* Fix formatting of if statement in itTomato_Logic9_Destroyed

* Fix bit field access for tomato item properties

* Implement pickup logic for Mario Cape item

Added logic for item pickup handling in itmariocape.c.

* Add includes for ftlib and it_2725 in itmariocape.c

Added new includes for ftlib and it_2725.

* Actualizar itmariocape.c

* Actualizar itmariocape.h

* Actualizar it_27CF.h

* Clean up itmariocape.c by removing blank lines

Removed unnecessary blank lines in itmariocape.c

* Add include for itmariocape in ftMr_SpecialS.c

* Add db.h include to itmariocape.c

* Fix function declaration for it_802B2560

* Implement itMariocape_UnkMotion1_Anim function

Add animation handling for Mario's cape item.

* Add declaration for itMariocape_UnkMotion1_Anim function

* Refactor it_802CCBF8 for improved readability

* Update x60 and x64 types in itThunder_ItemVars

Changed types of x60 and x64 from u32 to s32 in itThunder_ItemVars.

* Change u32 to f32 in itThunderPokemonAttributes

* Refactor itHeart_Logic8_Destroyed function parameters

* Change b0 type from u8 to s8 in itHeart_ItemVars

* Run `clang-format`

* Various cleanup

* Fix issues

---------

+++ src/melee/it/items/itlinkbomb.c
@@ -190,7 +190,7 @@ void it_8029DB5C(HSD_GObj* gobj)
 
 static inline void it_8029DD58_inline(Item* item, itLinkBombAttributes* attr,
                                       Item_GObj* fighter_gobj,
-                                      Item_GObj* temp_r3, u8 arg2)
+                                      Item_GObj* temp_r3, u32 arg2)
 {
     f32 temp_f2;
     it_80275158(temp_r3, attr->lifetime);
@@ -204,7 +204,7 @@ static inline void it_8029DD58_inline(Item* item, itLinkBombAttributes* attr,
     Item_8026AB54(temp_r3, fighter_gobj, arg2);
 }
 
-HSD_GObj* it_8029DD58(Item_GObj* fighter_gobj, Vec3* arg1, u8 arg2, int arg3,
+HSD_GObj* it_8029DD58(Item_GObj* fighter_gobj, Vec3* arg1, u32 arg2, int arg3,
                       f32 arg4)
 {
     Item_GObj* temp_r3;
